updated ex5

+++ tools/blocking.h
@@ -9,7 +9,7 @@
 
 namespace blocking
 {
-    double error(double av, double av_2, double n);
+    double error(double av, double av_2, int n);
     std::vector<std::vector<double>> prog_mean(std::function<double(Random &)> func, Random &rnd, int N, int L);
     void write_data(std::vector<std::vector<double>> data, std::string path, std::string header);
     void write_data(std::vector<double> data, std::string path, std::string header); // overloading of function above
